wrap native methods in secondary domain
